Make sure to use active local session in deployFromLocalActive
Move methods up to Session and generalize createSessionFromExisting()
