+++ configserver/src/main/java/com/yahoo/vespa/config/server/session/SessionFactoryImpl.java
@@ -118,8 +118,7 @@ public class SessionFactoryImpl implements SessionFactory, LocalSessionLoader {
         return session;
     }
 
-    @Override
-    public LocalSession createSessionFromExisting(RemoteSession existingSession,
+    public LocalSession createSessionFromExisting(Session existingSession,
                                                   DeployLogger logger,
                                                   boolean internalRedeploy,
                                                   TimeoutBudget timeoutBudget) {
