+++ configserver/src/test/java/com/yahoo/vespa/config/server/http/SessionHandlerTest.java
@@ -30,7 +30,6 @@ import com.yahoo.vespa.config.server.session.DummyTransaction;
 import com.yahoo.vespa.config.server.session.LocalSession;
 import com.yahoo.vespa.config.server.session.MockSessionZKClient;
 import com.yahoo.vespa.config.server.session.PrepareParams;
-import com.yahoo.vespa.config.server.session.RemoteSession;
 import com.yahoo.vespa.config.server.session.Session;
 import com.yahoo.vespa.config.server.session.SessionContext;
 import com.yahoo.vespa.config.server.session.SessionFactory;
@@ -215,8 +214,7 @@ public class SessionHandlerTest {
         }
 
         @Override
-        public LocalSession createSessionFromExisting(RemoteSession existingSession, DeployLogger logger,
-                                                      boolean internalRedeploy, TimeoutBudget timeoutBudget) {
+        public LocalSession createSessionFromExisting(Session existingSession, DeployLogger logger, boolean internalRedeploy, TimeoutBudget timeoutBudget) {
             if (doThrow) {
                 throw new RuntimeException("foo");
             }
